ADDED: Implementations FacetEdge

+++ CrystalMeshLib/Subdiv3/Subdiv3ForwardDecs.h
@@ -34,9 +34,7 @@ namespace CrystalMesh{
 		 * Define some types
 		 */
 
-		typedef uint8_t FieldIndex;
-
-
+		typedef int32_t FieldIndex;
 	}
 }
